part-4/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
loguru==0.7.2
python-dotenv==1.0.0
requests==2.31.0
selectolax==0.3.17
sentry-sdk==1.39.2
urllib3==2.1.0
